feat: Introduce CollapsibleActionMenu and PanelToggleButtons for enhanced HeadBar functionality and improved user experience in compact mode
